backend/requirements.txt
fastapi
uvicorn
cryptography
pydantic
python-multipart

# Speech-to-text
openai-whisper
torch

# NLP
transformers
scikit-learn


librosa
soundfile
numpy
